Painter: replace cairo_{fill,paint}() calls with libm2d API when possible
This is the first integration of drawing operations performed with the
hardware GPU when available, through libm2d.

Painter::fill() and Painter::paint() both tries to perform their
rendering operations with a new GPUPainter instance when HAVE_LIBM2D is
defined.

The GPUPainter class provides its own implementations of the fill() and
paint() methods, which execute calls to the relevant functions of libm2d
API.

If for some reason, GPUPainter::fill() or GPUPainter::paint() fail,
the Painter class falls back to calls of cairo_fill() or cairo_paint()
as before.

The GPUPainter class also has some sync_for_cpu() method to be called
when the CPU needs to claim the ownership of the target Surface and the
source Surface, if any, before accessing those surfaces with cairo or
anything else.

Finally, introduce a mean to globally disable the GPU (for benchmark
comparison between GPU and CPU rendering) either at startup by reading
the 'EGT_GPU_DISABLED' environment variable or at runtime with the
Application::gpu_enabled() getter and Application::enable_gpu() setter.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>