Slider: rework how sliders are drawn when images are provided
When no images are provided, sliders are drawn as before.

On the other hand, when images are provided for the slider background
or handle, change how they are drawn.

1 - A background image can be drawn now. Unlike what Widget::draw() does
with the background image, Slider::draw() doesn't resize it and places it
relatively to the center of the slider content area.

2 - The handle image is no longer resized to match the size of an handle
drawn without an image; instead the handle image now keeps its original
size. Also, the handle image is placed relatively to the center of the
background image, then shifted with the optional 'm_handle_offset'.

Moreover, the Slider::min_size_hint() method is reworked too, so both
the background and handle images can fit into the slider content area,
taking into account the 'm_handle_offset' and 'm_orient' values.

Finally, when an image is provided for the handle, the slider label, if
drawn, is now positioned relatively to the center of the handle image,
instead of being placed inside a dedicated region, but still taking into
account the 'm_label_offset' value.

All these changes allow to reproduce the design of the SliderB class
(from egt-sample-contribution) but drawing the slider components with
images rather than relying on cairo to draw curves.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Acked-by: Ludovic Desroches <ludovic.desroches@microchip.com>