Widget: add a new 'keep_ratio' AlignFlag
This new AlignFlag has precedence over the 'expand_horizontal' and
'expand_vertical' flags. When set, detail::align_algorithm_force()
expands the 'orig' rectangle to 'bounding' rectangle but preserving the
initial ratio between orig.width() and orig.height().

Especially, if a widget has this 'keep_ratio' flag set, when its parent
updates the layout, the parent resizes the widget so the child is
expanded to the parent content area but preserving ratio of the
'm_user_requested_box' member of the child.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>