Image: prevent the Image::load() method from throwing exceptions
The ImageCache::get() method may throw an exception, for instance when
its 'uri' argument is invalid or if it has failed to load the image.

Therefore, make the Image::load() method catch exceptions if any, then
reset the 'm_surface' and 'm_orig_size' members as if the 'uri' argument
were empty.

Also, log an error when an exception is caught.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>