Slider: fix Slider::draw_line() when the length of some line is zero
The documentation of the
'Painter& Painter::draw(const Color&, const RectF&, bool)' method states
that when the 'const RectF&' parameter is empty, the whole clip region
is filled with the color.

However, this is not the desired behavior when called from
Slider::draw_line(). Therefore, check whether rectangles are empty
before calling this Painter::draw() overload.

Fixes: 3333e6f49815 ("Slider: draw rectangles instead of lines in
Slider::draw_line()")
Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>