Slider: add the 'line_transition' property
The new 'line_transition' property controls the position of the
transition within the slider line between the 'normal' state and the
'disabled' state.

When set to 'to_end' (the default), the color(button_fg, normal) color
is selected to draw the line behind the slider handle.

When set the 'to_begin', the color(button_fg, disabled) color is
selected to draw the line behind the slider handle.

When set the 'to_center', the color(button_fg, normal) color is selected
to draw the first half of the line behind the slider handle whereas the
color(button_fg, disabled) color is used to draw the second half of the
line.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>