cairo_jpg: fix incompatible-pointer-types warning
This is the second attempt to fix the incompatible-pointer-types
warning. There are several implementations of the libjpeg and the
prototype for jpeg_mem_dest differs. With some implementations the last
parameter is a size_t pointer, with others it is an unsigned long
pointer.
Prefer a size_t* cast as it covers more cases.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>