example/camera: rework the overlays handling
This exemple requires several overlays: a high-end overlay for the
video, an overlay for the control window and another one for the dialog.
Some devices only provide on overlay in addition to the high-end one. As
a result, the dialog or control window is hidden by the overlays.
Change a bit the way overlays are used. Create a full screen window
overlay that handles both the controls and the dialog popup.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>