Image: don't assume strides are equal when copying an image into another
The strides of the source and destination images may be different;
especially if cairo (for the source) and the GPU (for the destination)
don't have the same alignment constraints for the stride.

For instance, cairo/pixman require the stride to be aligned to 4 bytes
whereas the Vivante GC520UL GPU requires the stride to be aligned to 16
bytes.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>
Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>