EventLoop: introduce a flush step to update the display
It is assumed that drawing handles the display update. Right now,
the display is updated for each window drawn. If we want to be able
to update the display only once per loop iteration, we need to split
the drawing and the screen update.

Let's introduce a new step called flush. Some graphics backends, as
the DRM one with the atomic API, can take benefit of it.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>