kmsscreen/overlay: override the flush() method
Libplanes 2.0.0 switched to the DRM Atomic API. This move change its
behavior. Instead of immediately committing the changes (framebuffer,
position, size of the plane, and others), they are gathered to be
committed all at once. The commit of the changes is triggered by
calling the new kms_device_flush function.

So override the flush() method to make it call kms_device_flush. Note
that the drm backend architecture should be reworked. Having a KMSScreen
and a KMSOverlay doesn't really makes sense. There are both subclasses
of Screen, but only one is really a screen.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>