KMSScreen: remove FlipJob and related code
The flip of the plane's framebuffer is handled by a FlipJob, that queue
the action. This queue is handled in a thread.

With the move to the DRM atomic API, the flip action is just an update
of the properties of the plane. The real flip happens when the changes
are committed, i.e. when the flush method is called. So we can get rid
of all this code and just call the plane_flip function.

Moreover, having the plane_flip in the thread may cause race issues. The
event loop may call the flush method before the thread updates the
properties of the planes.

The removal of flipthread causes some header inclusion issues. That's
why some includes have been added.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>