CMake: don't hard code the type of library to create
Add the BUILD_SHARED_LIBS option, and remove the type to add_library
to allow choosing SHARED or STATIC depending on the value of
BUILD_SHARED_LIBS.

For now, stick to this solution as build systems like Buildroot uses
this option. In the future, maybe we'll have to think about providing
a way to build both static and shared at once.

Signed-off-by: Ludovic Desroches <ludovic.desroches@microchip.com>
Acked-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>